+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 ## 0.0.15
 
 - Add default value for trait property when auto-completing from trait name
+- Provide description for trait completion in tasks.json
 
 ## 0.0.14